+++ src/main/java/com/stonytark/magnetization/registry/MagBlockEntities.java
@@ -96,6 +96,11 @@ public final class MagBlockEntities {
                     .of(com.stonytark.magnetization.content.sensor.MagnetostrictiveSensorBlockEntity::new, MagBlocks.MAGNETOSTRICTIVE_SENSOR.get())
                     .build(null));
 
+    public static final DeferredHolder<BlockEntityType<?>, BlockEntityType<com.stonytark.magnetization.content.induction.KineticCoilBlockEntity>> KINETIC_COIL =
+            REGISTER.register("kinetic_coil", () -> BlockEntityType.Builder
+                    .of(com.stonytark.magnetization.content.induction.KineticCoilBlockEntity::new, MagBlocks.KINETIC_COIL.get())
+                    .build(null));
+
     public static final DeferredHolder<BlockEntityType<?>, BlockEntityType<com.stonytark.magnetization.content.induction.InductionPadBlockEntity>> INDUCTION_PAD =
             REGISTER.register("induction_pad", () -> BlockEntityType.Builder
                     .of(com.stonytark.magnetization.content.induction.InductionPadBlockEntity::new, MagBlocks.INDUCTION_PAD.get())